% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uqfc.R
\name{empQU}
\alias{empQU}
\title{Extract historical empirical quantiles from WEO data}
\usage{
empQU(weodat, tv_release = 0.5, quantiles = c(0.1, 0.25, 0.5, 0.75, 0.9))
}
\arguments{
\item{weodat}{(subset of) WEO data as returned by download.process.weo()
function in package imfpp}

\item{tv_release}{which release of the true value in the WEO data shall be
used. Must be one of c(0.5, 1, 1.5, 2)}

\item{horizon}{horizon for which quantiles should be extracted}
}
\value{
data.table with empirical quantiles
}
\description{
Extract historical empirical quantiles from WEO data
}
