% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uqfc.R
\name{score_quants}
\alias{score_quants}
\title{Score quantile forecasts}
\usage{
score_quants(empquants, weodat, sRule = scoringRules::crps_sample)
}
\arguments{
\item{empquants}{data containing empirical quantiles}

\item{weodat}{(subset of) WEO data containing relevant truth data}
}
\value{
nothing, saves tidied data in directory
}
\description{
Score quantile forecasts
}
