% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uqfc.R
\name{scoreempQu}
\alias{scoreempQu}
\title{Score quantile forecasts}
\usage{
scoreempQu(fcdat, by = c("country", "target", "horizon"))
}
\arguments{
\item{fcdat}{forecast data, containing both true values and quantile predictions}
}
\value{
nothing, saves tidied data in directory
}
\description{
Score quantile forecasts
}
