% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uqfc.R
\name{year_set}
\alias{year_set}
\title{Produce year sets depending on respective estimation method}
\usage{
year_set(
  target_year,
  avail_years,
  method = c("leave-one-out", "rolling window", "expanding window"),
  window_length = NULL
)
}
\arguments{
\item{target_year}{year for which estimation set should be produced}

\item{avail_years}{all available years}

\item{method}{estimation method, one of 'leave-one-out', 'rolling window',
'expanding window'}
}
\value{
vector with years that should enter estimation set for given target_year
}
\description{
Produce year sets depending on respective estimation method
}
